Separating admin code from front end code to make it easier to add an ACL feature at a later date and to separate the concerns
