+++ app/Http/Controllers/Admin/GroupController.php
@@ -1,4 +1,4 @@
-<?php namespace App\Http\Controllers;
+<?php namespace App\Http\Controllers\Admin;
 
 use Lang;
 use App\Http\Requests;
